requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
diff-match-patch==20241021
distro==1.9.0
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.8
django-modeltranslation==0.19.16
django-select2==8.4.1
django-unfold==0.62.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8.2
openai==1.97.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.3.0
tablib==3.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
